requirements.txt
fastapi
mangum
uvicorn
pydantic
pydantic-settings
boto3
httpx
python-multipart
python-dotenv
